Fix TextBox line height type
\Intervention\Image\Typography\FontFactory lets you call `lineHeight`
with a float for more line height options. Currently, the package
only lets you provide a custom line height as an int.

Since the default value is 1.6, this fix lets you set it to e.g. 1.8,
which may be better if you are using a custom font or want to have
more granular control over the layout that is being rendered.

+++ src/Layout/TextBox.php
@@ -20,7 +20,7 @@ class TextBox extends Box
     public ColorInterface $color;
     public Font $font;
     public string $hAlign;
-    public int $lineHeight;
+    public float $lineHeight;
     public int $size;
     public string $text;
     public string $vAlign;
@@ -43,7 +43,7 @@ class TextBox extends Box
         return $this;
     }
 
-    public function lineHeight(int $lineHeight): self
+    public function lineHeight(float $lineHeight): self
     {
         $this->lineHeight = $lineHeight;
         return $this;
